moved auth stuff to auth folder
